Security config

+++ src/main/java/com/accio/book_my_show/Models/Ticket.java
@@ -35,7 +35,7 @@ public class Ticket {
 
     private String theaterAddress;
 
-    private Integer totalAmount;
+    private long totalAmount;
 
     //private List<String> bookedSeats;
